refactored the code to reduce codeclimate complexity; no more codeclimate issues
