+++ src/components/searchViewComponents/BaseList.vue
@@ -3,10 +3,14 @@ import CardContent from "@/components/searchViewComponents/subcomponents/generic
 import HoverCard from "./subcomponents/HoverCard.vue";
 import { TypeListNames } from "@/helpers/typeHelper";
 import { useFileListComposable } from "@/composables/useFileListComposable";
-import { onBeforeMount } from "vue";
+import { computed, onBeforeMount } from "vue";
 import { onBeforeRouteUpdate, useRoute } from "vue-router";
 import Hyperlink from "@/components/shared/HyperLink.vue";
+import { enterSearchQuery } from "@/router";
+import store from "@/store";
+import { batchSize, maxPages } from "@/helpers/ApiHelper";
 
+const infiniteScrollMargin = 200;
 const route = useRoute();
 
 const props = defineProps({
@@ -16,18 +20,88 @@ const props = defineProps({
   },
 });
 
-const {
-  resultsTotal,
-  anyFileType,
-  queryPage,
-  pageCount,
-  error,
-  loading,
-  infinite,
-  handleQueryChange,
-  slicedHits,
-  getResultsOnMount,
-} = useFileListComposable(props);
+const { anyFileType, loading, loadedPages, infinite, slicedHits } = useFileListComposable(props);
+
+const error = computed(() => {
+  const error = store.getters[`results/${props.fileType}/error`];
+  if (error) console.error("Error loading results", error);
+  return error;
+});
+
+const resultsTotal = computed(() => {
+  const resultsTotalMax = 10000;
+  const total = store.getters[`results/${props.fileType}/resultsTotal`];
+  if (total === resultsTotalMax) {
+    return "10000+";
+  }
+  return total;
+});
+
+const pageCount = computed(
+  () => Math.min(Math.ceil(store.getters[`results/${props.fileType}/resultsTotal`] / batchSize)),
+  maxPages
+);
+
+/**
+ * See if the the page scrolled so far down that empty space opens up at the bottom.
+ * Also update the url
+ * used by ImageList
+ */
+const infiniteScroll = () => {
+  if (!infinite.value) return;
+  const { scrollTop, scrollHeight } = document.documentElement;
+  // calculate, which page is currently in view
+  const scrollPage = Math.floor(loadedPages.value * (scrollTop / scrollHeight)) + 1;
+  // if needed, change the page in the URL
+  if (store.state.query.page !== scrollPage) {
+    enterSearchQuery(route.query, scrollPage, "replace");
+  }
+  const nearBottom = window.innerHeight + infiniteScrollMargin > scrollHeight - scrollTop;
+  if (nearBottom && !loading.value) {
+    return store.dispatch(`results/${props.fileType}/fetchPage`, { page: loadedPages.value + 1 });
+  }
+};
+
+/**
+ * scroll down to the page from the query
+ */
+const scrollDown = () => {
+  const scrollQueryPage = Number(route.query.page);
+  const { scrollHeight } = document.documentElement;
+  let top = 0;
+  if (scrollQueryPage > 1) {
+    top = ((scrollHeight - window.innerHeight) / loadedPages.value) * (scrollQueryPage + 1);
+  }
+  window.scrollTo({
+    top,
+    left: 0,
+    behavior: "smooth",
+  });
+};
+
+function handleQueryChange(query = route.query) {
+  if (infinite.value) {
+    // if (query.type === fileType) {
+    //   document.addEventListener('scroll', infiniteScroll, true);
+    // } else {
+    //   document.removeEventListener('scroll', infiniteScroll, true);
+    // }
+    return Promise.all(
+      Array.from({ length: Math.min(store.state.query.page + 1, maxPages) }, (_, i) =>
+        store.dispatch(`results/${props.fileType}/fetchPage`, { page: i + 1 })
+      )
+    );
+  }
+  return store.dispatch(`results/${props.fileType}/fetchPage`, { page: query.page || 1 });
+}
+
+function getResultsOnMount() {
+  if (infinite.value) {
+    handleQueryChange().then(scrollDown).then(infiniteScroll);
+  } else {
+    handleQueryChange();
+  }
+}
 
 onBeforeMount(getResultsOnMount);
 
@@ -36,10 +110,19 @@ onBeforeRouteUpdate(({ query }, from, next) => {
   handleQueryChange(query);
   next();
 });
+
+const queryPage = computed({
+  get() {
+    return Number(route.query.page);
+  },
+  set(value) {
+    enterSearchQuery(route.query, value);
+  },
+});
 </script>
 
 <template>
-  <v-container class="overflow-y-auto h-100">
+  <v-container v-scroll="infiniteScroll" class="overflow-y-auto h-100">
     <v-row dense>
       <v-col cols="12" xl="8" offset-xl="2">
         <v-card flat>
